fix: correctly expose gen_test_data cxx ffi according to VORTEX_ENABLE_TESTING
