+++ vortex-cxx/examples/README.md
@@ -1,0 +1,11 @@
+# C++ examples
+
+This example shows how to interface with the C++ API using CMake.
+
+```bash
+mkdir -p build
+cd build
+cmake ..
+make -j
+./hello-vortex ../example.vortex
+```
